requirements.txt
requests
pydantic
python-dotenv
python-telegram-bot
